Show reset_resp in wrong YAML, no nesting actions
Trying to start the resend activation action before the registration
action has technically finished (on the last screen, but DA still
needs that continue variable) queues up the next action, but stays
on the current screen, seeming like nothing is happening. Instead,
don't show that action as an option, just push it to the next screen.