Added the Visible Password functionality
Adds the Visible Password custom datatype, that lets you view your password
when entering it. It doesn't change anything about the datatype on the backend,
just some additional javascript on the frontend (a checkbox to show the
password).

Additionally, shows Tyler's error messages when trying to reset passwords, which
do say if the email exists or not.