Use intended way to see if people are individuals
Strangely, to see if a user is an individual vs a part of a firm, you have
to call "getFirm()" which all users have (it's where the address is for
individuals) and check the `isIndividual` attribute. This should reduce
the number of false positives Tyler Error logs that we trigger on the proxy
server, specifically "Operation not available for individual filers".