Catch when no payment methods (#233)
Before, if the user didn't have any payment methods added, we'd silently skip it and then eventually error. This PR adds back a screen that asks users to make a new payment method. It refreshes when the user adds new methods as well.

Also corrects a grammar mistake in that payment flow.