Fix attribute issue
Forgot that in docassemle, you can't assume your objects will be correctly initialized,
and should do it on the fly.

Fixed this issue for `session_id` and `logger`, the two attributes that were added in
#274. Have to handle the session ID in just the parent class (`EfspConnection`), but
need to handle the logger in both the parent and child class (`ProxyConnection`).

(Have to have the methods in the parent class at all because the child class because
the parent class would have the same issues. This makes separating out a full python
client more complicated than I realized unfortunately

Reintegrating myself into the world of DA, one fix at a time).