Don't depend on `declaredType'
Directly couples us to the full java package namespace, which is not good
(there have been refactors that are out in efile-test now that break the
assumption that the package names won't change).

Instead, use `name` (which is the XML Schema name), or just
check that the `declaredType` contains the path we expect (those
refactors only add to the namespace, not changing it entirely).